Replace findPermit with filterPermit
