+++ packages/gafl-webapp-service/src/processors/find-permit.js
@@ -1,45 +1,36 @@
-import filterPermits from './filter-permits.js'
-import crypto from 'crypto'
-import db from 'debug'
-const debug = db('webapp:find-permit')
+import { salesApi } from '@defra-fish/connectors-lib'
 
-export const findPermit = async (permission, request) => {
-  /*
-   * To stop repeated reads of the API with users repeatably refreshing the page, the transaction cache stores
-   * a hash of itself. If the transaction cache has not changed the permit is not recalculated.
-   *
-   * The section of the transaction cache subject to the hashing algorithm excludes
-   * name, address, or anything not effecting permit filter
-   */
-  const hashOperand = (({ hash: _hash, permit: _permit, licensee: _licensee, ...p }) => p)(permission)
+export const getPermitsJoinPermitConcessions = async () => {
+  const permits = await salesApi.permits.getAll()
+  const permitConcessions = await salesApi.permitConcessions.getAll()
+  const concessions = await salesApi.concessions.getAll()
+  return permits.map(p => ({
+    ...p,
+    concessions: permitConcessions.filter(pc => pc.permitId === p.id).map(pc => concessions.find(c => c.id === pc.concessionId))
+  }))
+}
 
-  // To calculate a permit, hash and save
-  const addHashAndPermit = async () => {
-    const permit = await filterPermits(permission)
-    permission.permit = permit
-    if (permit) {
-      if (!permit.newCostStartDate || !permit.newCost) {
-        debug('permit missing new cost details', permission)
-      }
-    } else {
-      debug("permit wasn't retrieved", permission)
-    }
+export default async permission => {
+  const licenseeConcessions = permission.concessions || []
+  const permitsJoinPermitConcessions = await getPermitsJoinPermitConcessions()
 
-    const hash = crypto.createHash('sha256')
-    hash.update(JSON.stringify(hashOperand))
-    permission.hash = hash.digest('hex')
-    await request.cache().helpers.transaction.setCurrentPermission(permission)
-  }
+  // Filter the joined list to include every and only those concessions in licenseeConcessions
+  const filteredPermitsJoinPermitConcessions = permitsJoinPermitConcessions.filter(
+    pjpc =>
+      licenseeConcessions.map(lc => lc.type).every(t => pjpc.concessions.map(c => c.name).includes(t)) &&
+      pjpc.concessions.length === licenseeConcessions.length
+  )
 
-  if (!permission.hash) {
-    await addHashAndPermit()
-  } else {
-    const hash = crypto.createHash('sha256')
-    hash.update(JSON.stringify(hashOperand))
-    if (hash.digest('hex') !== permission.hash) {
-      await addHashAndPermit()
-    } else {
-      debug("permit data present and doesn't need updating")
-    }
-  }
+  // Filter by the licence length
+  const byLicenceLength = filteredPermitsJoinPermitConcessions.filter(
+    p => String(p.durationMagnitude + p.durationDesignator.description) === permission.licenceLength
+  )
+
+  // Filter by the licence (sub) type
+  const byLicenceType = byLicenceLength.filter(p => p.permitSubtype.label === permission.licenceType)
+
+  // Filter by the number of rods
+  const byNumberOfRods = byLicenceType.filter(r => String(r.numberOfRods) === permission.numberOfRods)
+
+  return byNumberOfRods[0]
 }
